Upload a bunch of games in the new (wip) format
This'll allow me to test the backend, which will be the most important part. The only stuff currently done rules.txt, plus 1 ultrawide patch. It's all for testing for now and it's pretty likely I'll add more stuff. I also skipped a few of the larger games as I want to make sure that everything besides shaders work good before all this. The current ultrawide 'placeholder' isn't very done yet. I also need to figure out if I need to clarify it some more, but I think it would be fine to assume that all patches.txt next to the resolution rules.txt are ultrawide patches. I want to fetch the files as late as possible due to performance. The actual 'building' time isn't very important if affected.