Reworking code
Most of the actual presentation is gone, but it's handled a lot better. Next step will be to have the options be parsed to some saveable state, which can dynamically be changed by the user if you turn back. Also need a way to figure out how to combine the 'advanced' options with some users that just skip to the download after selecting their games.