WIP syntax rules, clear 'demo' code
Cleaned the code, as the previous thing really was to test the whole GUI and flow of the design a bit.

The syntax that I think I'm gonna stick with is very wip, but I think it'll make all of the use cases the most convenient and simple to use.