Split Javascript from HTML
Due to Github CDN preventions I thought I had to include the script, but I got mixed up and recently found that I mixed that up.

I've also commented/changed a small part of the code which is now explains stuff somewhat clearer.

There's 1 major thing I've added, which is the compat fetch code which is now somewhat working, although it seems that the regex is causing issues.