Minor changes to various elements
Mostly bothered with changes that would prevent me getting errors about guidelines.

The localstorage warning is also in the footer now, which it should've been in to begin with.

I've been using Visual Studio for a bit too. I really want to clean up some of the css, as I think it's the messiest part (I'm not too familiar with it either), so I think I'll start working on the Settings panel and add the CSS rules for a dark theme/themes.