Themes and settings!
Okay, so the whole thing is now not only for this repository, but can be forked or build from a different repository as it's now an option to set your own repository. This'll be used instead of the previously where I had this github repository hardcoded in. It'll use the github pages url to get the default, but for e.g. local building/forking you will have to do change that.

The theme option is super not finished, especially since I haven't had the time to tweak the dark theme to something I'd like. But it's there in the settings menu! There's also default values now for that. The css also uses css properties to tweak colors, which makes it somewhat more readable and had me simplify the color palette a bit. The settings are also better stored, somewhat.