Custom themes! Improved code paths
Theme's are properly implemented and save and restore properly instead of the demo code that was included before. The CSS theme system is also added in pretty much all of the places and look good in dark and light mode.

Custom theme support has sharing functions for exchanging themes and a color picker menu.

The code paths (mainly the button hover from the game's selection) have also been improved due to simplifying the animations to transitions and moving the code to CSS. Squashed some bugs too with the settings popup.