Small fixes/additions
Just some small fixes that I might overlook if I revert to the previous commit.

Planning to finally do the big work, but I've decided I want to change the order/position where the parsing will happen. The current parsing is using local methods, and I'd only consider using those for the main rules.txt files that contain the metadata. That whole parsing method will have to be redesigned to fit/apply to the other rules.txt, shader files or patches, which wouldn't be good design either. Plus, it's better to not execute that whole parsing method for the about 70 packs we'll have at the beginning. Besides the parsing, it'll likely affect some other things and change the usage in the various files.